+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Administration/Warn.cs
@@ -2,7 +2,6 @@
 using Discord.Commands;
 using Discord.WebSocket;
 using KaguyaProjectV2.KaguyaBot.Core.Attributes;
-using KaguyaProjectV2.KaguyaBot.Core.Handlers;
 using KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent;
 using KaguyaProjectV2.KaguyaBot.Core.KaguyaEmbed;
 using KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Models;
@@ -61,20 +60,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Administration
             }
 
             await ReplyAsync(embed: (await Reply(wu, user)).Build());
-
-            if (server.IsPremium && server.ModLog != 0)
-            {
-                var premLog = new PremiumModerationLog
-                {
-                    Server = server,
-                    Moderator = (SocketGuildUser) Context.User,
-                    ActionRecipient = user,
-                    Reason = reason,
-                    Action = PremiumModActionHandler.WARN
-                };
-
-                await PremiumModerationLog.SendModerationLog(premLog);
-            }
 
             await DatabaseQueries.UpdateAsync(server);
         }