+++ KaguyaProjectV2/KaguyaBot/Core/Handlers/WarnEvent/WarnHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Discord.WebSocket;
 using KaguyaProjectV2.KaguyaBot.Core.Commands.Administration;
 using KaguyaProjectV2.KaguyaBot.Core.Global;
@@ -12,7 +13,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
 {
     public static class WarnHandler
     {
-        public static async void OnWarn(object warn, WarnHandlerEventArgs args)
+        public static async Task OnWarn(WarnHandlerEventArgs args)
         {
             var currentSettings = await DatabaseQueries.GetFirstForServerAsync<WarnSetting>(args.Server.ServerId);
 
@@ -35,14 +36,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
             if (warnCount == banNum)
             {
                 var ban = new Ban();
-                var modLog = new PremiumModerationLog
-                {
-                    Server = args.Server,
-                    Moderator = kaguya,
-                    ActionRecipient = guildUser,
-                    Action = PremiumModActionHandler.AUTOBAN,
-                    Reason = $"Automatic ban due to the user reaching {warnCount} warnings."
-                };
 
                 try
                 {
@@ -50,7 +43,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
                                                           $"reaching the specified warning threshold for bans " +
                                                           $"({warnCount} warnings).");
 
-                    await PremiumModerationLog.SendModerationLog(modLog);
                     await ConsoleLogger.LogAsync($"User [{guildUser} | {guildUser.Id}] has been " +
                                                  $"automatically banned in guild " +
                                                  $"[{guildUser.Guild} | {guildUser.Guild.Id}]", LogLvl.DEBUG);
@@ -71,19 +63,10 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
             if (warnCount == shadowbanNum)
             {
                 var shadowban = new Shadowban();
-                var modLog = new PremiumModerationLog
-                {
-                    Server = args.Server,
-                    Moderator = kaguya,
-                    ActionRecipient = guildUser,
-                    Action = PremiumModActionHandler.AUTOSHADOWBAN,
-                    Reason = $"Automatic shadowban due to the user reaching {warnCount} warnings."
-                };
 
                 try
                 {
                     await shadowban.AutoShadowbanUserAsync(guildUser);
-                    await PremiumModerationLog.SendModerationLog(modLog);
                     await ConsoleLogger.LogAsync($"User [{guildUser} | {guildUser.Id}] has been " +
                                                  $"automatically shadowbanned in guild " +
                                                  $"[{guildUser.Guild} | {guildUser.Guild.Id}]", LogLvl.DEBUG);
@@ -104,14 +87,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
             if (warnCount == kickNum)
             {
                 var kick = new Kick();
-                var modLog = new PremiumModerationLog
-                {
-                    Server = args.Server,
-                    Moderator = kaguya,
-                    ActionRecipient = guildUser,
-                    Action = PremiumModActionHandler.AUTOKICK,
-                    Reason = $"Automatic kick due to the user reaching {warnCount} warnings."
-                };
 
                 try
                 {
@@ -119,7 +94,6 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
                                                             $"reaching the specified warning threshold for kicks " +
                                                             $"({warnCount} warnings).");
 
-                    await PremiumModerationLog.SendModerationLog(modLog);
                     await ConsoleLogger.LogAsync($"User [{guildUser} | {guildUser.Id}] has been " +
                                                  $"automatically kicked in guild " +
                                                  $"[{guildUser.Guild} | {guildUser.Guild.Id}]", LogLvl.DEBUG);
@@ -140,19 +114,10 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Handlers.WarnEvent
             if (warnCount == muteNum)
             {
                 var mute = new Mute();
-                var modLog = new PremiumModerationLog
-                {
-                    Server = args.Server,
-                    Moderator = kaguya,
-                    ActionRecipient = guildUser,
-                    Action = PremiumModActionHandler.AUTOMUTE,
-                    Reason = $"Automatic mute due to the user reaching {warnCount} warnings."
-                };
-
+                
                 try
                 {
                     await mute.AutoMute(guildUser);
-                    await PremiumModerationLog.SendModerationLog(modLog);
                     await ConsoleLogger.LogAsync($"User [{guildUser} | {guildUser.Id}] has been " +
                                                  $"automatically muted in guild " +
                                                  $"[{guildUser.Guild} | {guildUser.Guild.Id}]", LogLvl.DEBUG);